fix: cast v_attnum::smallint in 00084/00085 for pg_constraint conkey comparison

+++ supabase/migrations/00084_bugfixes_stock_checkout.sql
@@ -235,7 +235,7 @@ BEGIN
     WHERE conrelid = 'public.stock_movements'::regclass
       AND contype = 'f'
       AND confrelid = 'public.inventory'::regclass
-      AND conkey = ARRAY[v_attnum];
+      AND conkey = ARRAY[v_attnum::smallint];
     IF v_conname IS NOT NULL THEN
       EXECUTE format('ALTER TABLE public.stock_movements DROP CONSTRAINT %I', v_conname);
     END IF;
